chore: Rename seacher.py to search.py
