+++ tests/test_search.py
@@ -1,7 +1,7 @@
 import tensorflow as tf
 
 from speech_recognition.model import LAS
-from speech_recognition.searcher import Searcher
+from speech_recognition.search import Searcher
 
 
 def test_search():
